Actually it's not a singleton, it's a wrapper around a JupyterServer. Good thing we have functional tests.
